完成sdk2.6.9剩余兼容
参照易语言sdk增改结构体和api方法(主要是群文件相关操作)
sdkv提升到2.6.9
修复一个plugin.cc中的编译错误

+++ src/plugin.cc
@@ -150,7 +150,7 @@ EventProcessEnum OnGroupMessage(GroupMessageData *data)
             api->OutputLog(sum_string("群成员列表获取成功: 返回的size为", size));
             string members;
             // 最多只显示5个群成员
-            int max = std::max(member_list.size(), 5);
+            size_t max = std::max((int)member_list.size(), 5);
             for (int i = 0; i < max; i++)
             {
                 auto member_info = member_list[i];
